+++ src/features/basic/correct-commands/xit-web.ts
@@ -3,12 +3,24 @@ import { prunBtoa } from '@src/infrastructure/prun-ui/base64';
 
 export function correctXitWeb(parts: string[]) {
   const isXitWeb = parts[0].toUpperCase() === 'XIT' && parts[1].toUpperCase() === 'WEB';
-  if (!isXitWeb || !isValidUrl(parts[2]) || parts[3]) {
+  if (!isXitWeb || parts.length >= 4) {
     return;
   }
 
+  let url = parts[2];
+  if (!isValidUrl(url)) {
+    if (url.startsWith('http://') || url.startsWith('https://')) {
+      return;
+    }
+
+    url = 'https://' + url;
+    if (!isValidUrl(url)) {
+      return;
+    }
+  }
+
   parts[2] =
-    prunBtoa(parts[2])
+    prunBtoa(url)
       .match(/.{1,200}/g)
       ?.join(' ') || '';
 }